gau gau

+++ src/Redux/Action/user.js
@@ -1,22 +1,35 @@
 
-import Swal from 'sweetalert2'
+import Swal from 'sweetalert2';
 import { createAction } from '.';
-import { UserServices } from '../../Services';
-import { THONGTINTAIKHOAN } from './type';
+import { ProductsService, UserServices } from '../../Services';
+import { DANHSACHUSER, THONGTINTAIKHOAN } from './type';
 export const DangNhap = (data) => {
 
     UserServices.dangNhap(data).then(res => {
-        console.log(res.data);
-        Swal.fire({
-            position: 'center',
-            icon: 'success',
-            title: res.data.message,
-            showConfirmButton: false,
-            timer: 1500
-        });
-        const user = JSON.stringify(res.data.data)
-        console.log("user>", user);
-        localStorage.setItem("user", user)
+
+        if (res.data.status == false) {
+            Swal.fire({
+                position: 'center',
+                icon: 'error',
+                title: "Sai mật khẩu",
+                showConfirmButton: false,
+                timer: 1200
+            });
+        } else {
+            console.log(res.data);
+            Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: "Đăng nhập thành công ! ",
+                showConfirmButton: false,
+                timer: 1500
+            });
+            const user = JSON.stringify(res.data.data)
+            console.log("user>", user);
+            localStorage.setItem("user", user)
+            window.location.reload(false);
+
+        }
     }).catch(err => {
         console.log(err);
         Swal.fire({
@@ -30,29 +43,7 @@ export const DangNhap = (data) => {
 
 }
 export const CapNhatThongTin = (token, data, id) => {
-    UserServices.capNhatThongTin(token, data, id).then(res => {
-        console.log(res.data, "thanh cong");
-        Swal.fire({
-            position: 'center',
-            icon: 'success',
-            title: 'Cập nhật thông tin thành công ! ',
-            showConfirmButton: false,
-            timer: 1500
-        });
-        localStorage.removeItem("user")
-        window.location.reload(false);
 
-
-    }).catch(err => {
-        console.log(err);
-        Swal.fire({
-            position: 'center',
-            icon: 'error',
-            title: err.response.data.message,
-            showConfirmButton: false,
-            timer: 1200
-        });
-    })
 
 }
 export const ThongTinTaiKhoan = (token) => {
@@ -61,7 +52,18 @@ export const ThongTinTaiKhoan = (token) => {
             dispatch(createAction(THONGTINTAIKHOAN, res.data.data.user))
             console.log("thongtinthanhcong");
         }).catch(err => {
-            console.log("Errr",err);
+            console.log("Errr", err);
+        })
+    }
+}
+
+export const DanhSachNguoiDung = ()=>{
+    return dispatch=>{
+        UserServices.danhSachNguoiDung().then(res=>{
+            console.log("danhsachuser",res.data.data.users);
+            dispatch(createAction(DANHSACHUSER,res.data.data.users))
+        }).catch(err=>{
+            console.log(err);
         })
     }
 }
@@ -81,9 +83,10 @@ export const DangKy = (data) => {
     }).catch(err => {
         Swal.fire({
             position: 'center',
+            showConfirmButton: false,
             icon: 'error',
             title: err.response.data.message,
-            showConfirmButton: false,
+
             timer: 1200
         });
     })
